Distinguish btwn garden & navbar search results
