+++ js/react/src/taxa/search.jsx
@@ -13,32 +13,60 @@ class TaxaSearchResults extends React.Component {
     this.state = {
       viewType: "grid"
     };
+
+    this.getPageHeader = this.getPageHeader.bind(this);
+  }
+
+  getPageHeader() {
+    if (this.props.family) {
+      return <h1 style={{ display: this.props.family !== null ? "initial" : "none"  }}>Search results for the { this.props.family } family</h1>;
+
+    } else if (this.props.genus) {
+      return <h1 style={{ display: this.props.genus !== null ? "initial" : "none" }}>Search results for the { this.props.genus } genus</h1>;
+
+    } else if (this.props.results.length >= 1) {
+      return (
+        <h1 style={{ display: this.props.results.length > 1 ? "intial" : "none" }}>
+          Results for "{ this.props.searchText }"
+        </h1>
+      );
+    } else {
+      return (
+        <div>
+          <h1>
+            Whoops, we didn't find any results for "{ this.props.searchText }"
+          </h1>
+          <button className="btn btn-primary my-4" onClick={ () => window.history.back() }>Go back</button>
+        </div>
+      );
+    }
   }
 
   render() {
     return (
-      <div className="mx-auto my-5 py-2" style={{ maxWidth: "75%" }}>
-        <h1 style={{ display: this.props.family !== null ? "initial" : "none"  }}>Search results for the { this.props.family } family</h1>
-        <h1 style={{ display: this.props.genus !== null ? "intial" : "none" }}>Search results for the { this.props.genus } genus</h1>
-        <SearchResultContainer viewType={ this.state.viewType }>
-          {
-            this.props.results.map((result) => {
-              if (result.images.length > 0) {
-                return (
-                  <SearchResult
-                    key={result.tid}
-                    viewType="grid"
-                    display={true}
-                    href={ getTaxaPage(CLIENT_ROOT, result.tid) }
-                    src={ result.images[0].thumbnailurl }
-                    commonName={ getCommonNameStr(result) }
-                    sciName={ result.sciname ? result.sciname : '' }
-                  />
-                );
-              }
-            })
-          }
-        </SearchResultContainer>
+      <div className="mx-auto my-5 py-3" style={{ maxWidth: "75%" }}>
+        { this.getPageHeader() }
+        <div style={{ minHeight: "30em" }}>
+          <SearchResultContainer viewType={ this.state.viewType }>
+            {
+              this.props.results.map((result) => {
+                if (result.images.length > 0) {
+                  return (
+                    <SearchResult
+                      key={result.tid}
+                      viewType="grid"
+                      display={true}
+                      href={ getTaxaPage(CLIENT_ROOT, result.tid) }
+                      src={ result.images[0].thumbnailurl }
+                      commonName={ getCommonNameStr(result) }
+                      sciName={ result.sciname ? result.sciname : '' }
+                    />
+                  );
+                }
+              })
+            }
+          </SearchResultContainer>
+        </div>
       </div>
     );
   }
@@ -47,7 +75,8 @@ class TaxaSearchResults extends React.Component {
 TaxaSearchResults.defaultProps = {
   results: [],
   family: null,
-  genus: null
+  genus: null,
+  searchText: ""
 };
 
 const domContainer = document.getElementById("react-taxa-search-app");
@@ -59,12 +88,8 @@ if (queryParams.search) {
     if (res.length === 1) {
       window.location = `./index.php?taxon=${res[0].tid}`
 
-    } else if (res.length > 1) {
-      ReactDOM.render(<TaxaSearchResults results={ res } />, domContainer);
-
     } else {
-      window.location = "/";
-
+      ReactDOM.render(<TaxaSearchResults results={ res } searchText={ decodeURIComponent(queryParams.search) } />, domContainer);
     }
   }).catch((err) => {
     console.error(err);
